Integrate GZA into cuckoo
Changes were made to GZA to make integration easier. Specifically:

* packets are altered by scapy, but are now delivered by nfqueue. scapy was
failing to send packets faithfully due to changes in the networking structure
* cuckoo makes use of options["custom"] to identify the game to be played.
this is a hack that should be replaced by its own option (--game) but is good
enough for now

Care MUST be made to ensure that this can be done with multiple VMs cycling in
and out.