Generate pool of nfqueues; use through whole run.
This represents a substantial departure from how GZA was structured in earlier
versions. The main difference is numvm * 2 nfqueues are pre-allocated, the
first half are for dnsw games and the second half are for tcpw games (since
these are the ones used most frequently). The nfqueues are initialized in a
separate forked off process which are terminated after the main cuckoo process
exits.