Fix snapshots when they lose --current status
If there's a serious bug in cuckoo during execution, the VM being used can lose
track of which snapshot is current. This prevents it from being used in future
runs until it is fixed. This script checks for the lack of a --current snapshot
and sets the latest snapshot to --current if it is missing.